refactor stats module
