--- iapyx/src/stats/harvester.rs
@@ -1,15 +1,0 @@
-pub struct Harvester {
-    address: String
-}
-
-impl Harvester {
-
-}
-
-
-pub struct Snapshot {
-    timestamp: Date,
-    votes_count: u64,
-    pending: u64,
-    total_tx: u64
-}
