--- iapyx/src/stats/settings.rs
@@ -1,4 +1,0 @@
-pub struct Settings {
-    interval: u64,
-    logger: PathBuf, 
-}
